LayoutTests/svg/animations/animate-text-nested-transforms-expected.txt

Test nested text transforms with animateMotion

This is some text.
test to determine whether auto-rotate animateMotion path animations pre-muliply or post-multiply animation transform matrix

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



    
  

PASS rootSVGElement.getBBox().x is almost 390 (within 20)
PASS rootSVGElement.getBBox().y is almost 140 (within 20)
PASS rootSVGElement.getBBox().x is almost 700 (within 20)
PASS rootSVGElement.getBBox().y is almost 270 (within 20)
PASS successfullyParsed is true

TEST COMPLETE
